Restore ability to specify revs and paths on the command line.
On Lion, some of the document opening logic is done in a block and
requeued on the main dispatch queue. During this time Apple Events are
suspended.

This means that getting the currentAppleEvent won't work, so the
revlists weren't being used in the display.

The solution was to override
openDocumentWithContentsOfURL:display:completionHandler, and save a copy
of the Apple Event in a dictionary. showWindows: looks there if there
isn't a current apple event.

This seems to work well, except on initial launch.